Alterações de gerenciamento de Bebidas
Refatoração de métodos que gerenciam os bancos de Bebidas em uma única classe...

+++ Projeto_PDSI_II/src/Alterações/Alteracao_Estoque_Bebidas.java
@@ -275,7 +275,7 @@ public class Alteracao_Estoque_Bebidas extends JFrame implements ActionListener 
             stmt.setString(3, dados_bebida.getFornecedor());
             stmt.setString(4, dados_bebida.getVolume());
             stmt.setInt(5,dados_bebida.getQuantidade());
-            stmt.setFloat(6, dados_bebida.getPreco());
+            stmt.setDouble(6, dados_bebida.getPreco());
                                  
             stmt.executeUpdate();
